GameForge-Django/GameForge-Django/requirements.txt
Django==4.2.15
Pillow==10.3.0
python-dotenv==1.0.1
requests==2.32.3
reportlab==4.2.2
